+++ public/admin/_app/immutable/entry/start.D1IXo3n3.js
@@ -1,0 +1,1 @@
+import{n as e,s as t}from"../chunks/DSF8-oBo.js";export{t as load_css,e as start};